% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{acceptance_ratio}
\alias{acceptance_ratio}
\title{Calculate Acceptance Ratio for Metropolis Parameters}
\usage{
acceptance_ratio(mod, params = c("all", "theta", "rho"), burn = 0)
}
\arguments{
\item{mod}{a model object created with the \code{init()} function.}

\item{params}{a string vector of parameters to load samples for. \code{all} specifies \code{theta} and \code{rho}.}

\item{burn}{amount of burn-in to be done on samples. Discards first \code{burn} iterations of the samples.}
}
\value{
A list containing the acceptance ratios for each \code{theta} and/or \code{rho} parameter.
}
\description{
Calculate Acceptance Ratio for Metropolis Parameters
}
\examples{
\dontrun{
acceptance_ratio(mod_nc)
acceptance_ratio(mod_nc, params = "all")
acceptance_ratio(mod = mod_nc, params = "theta", burn = 2e3)
}
}
