% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxfunc.R
\name{get_medians}
\alias{get_medians}
\title{Calculate median estimates from samples}
\usage{
get_medians(rec_samples, params = c("all", names(rec_samples)), ci = 0.95)
}
\arguments{
\item{rec_samples}{a list of samples created with the \code{load_samples()} function.}

\item{params}{a string vector of parameters to load samples for. \code{all} specifies all parameters.}
}
\value{
A list containing median estimates for each parameter of interest.
}
\description{
Calculate median estimates from samples
}
\examples{
\code{load_samples(output_nc)}
\code{load_samples(output_nc, params = "all")}
\code{load_samples(output_nc, c("theta", "tau2", "Gt"))}
}
