% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{get_reliability}
\alias{get_reliability}
\title{Calculate reliability of median estimates}
\usage{
get_reliability(medians, params = c("all", names(medians)))
}
\arguments{
\item{medians}{a list of medians created with the \code{get_medians()} function.}

\item{params}{a string vector of parameters to load samples for. \code{all} specifies all parameters.}
}
\value{
A list containing median estimates for each parameter of interest.
}
\description{
Calculate reliability of median estimates
}
\examples{
\dontrun{
get_reliability(medians_nc)
get_reliability(medians_nc, c("theta", "tau2", "Gt"))
}
}
