% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_sampler}
\alias{gibbs_sampler}
\title{Gibbs Sampler Updates}
\usage{
gibbs_sampler(mod, n_iter, n_loop = 0L, l = 0L)
}
\arguments{
\item{mod}{List of model including model data, neighbor data,
hyperparameters, and output}

\item{n_iter}{number of iterations to append}
}
\value{
A list containing model data, neighbor data, hyperparameters,
and initial output
}
\description{
Gibbs Sampler Updates
}
