% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{load_model}
\alias{load_model}
\title{Load Model From Directory}
\usage{
load_model(name, dir = getwd())
}
\arguments{
\item{name}{the name assigned to the model in the \code{init} function.}

\item{dir}{the directory where the model exists.}
}
\value{
A list containing model data, neighbor data, hyperparameters,
and initial output
}
\description{
Load Model From Directory
}
\examples{
\dontrun{
load_model("heart_nc")
load_model(name = "heart_nc", dir = getwd())
}
}
