% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_functions.R
\name{load_samples}
\alias{load_samples}
\title{Load Samples From Storage}
\usage{
load_samples(mod, params = c("all", names(mod$inits)), thin = 1, burn = 0)
}
\arguments{
\item{mod}{a model object created with the \code{init()} function.}

\item{params}{a string vector of parameters to load samples for. \code{all} specifies all parameters.}

\item{thin}{amount of thinning to be done on samples. Loads in every \code{thin} iteration of the samples.
Must multiply evenly into 100.}

\item{burn}{amount of burn-in to be done on samples. Discards first \code{burn} iterations of the samples.}
}
\value{
A list containing samples for each parameter of interest.
}
\description{
Load Samples From Storage
}
\examples{
\dontrun{
load_samples(mod_nc)
load_samples(mod_nc, params = "all")
load_samples(mod = mod_nc, params = "theta", thin = 10, burn = 2e3)
load_samples(mod_nc, c("theta", "tau2", "Gt"), burn = 2e3)
}
}
