\name{mstcar-package}
\alias{mstcar-package}
\alias{mstcar}
\docType{package}
\title{
\packageTitle{mstcar}
}
\description{
\packageDescription{mstcar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mstcar}
\packageIndices{mstcar}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mstcar}

Maintainer: \packageMaintainer{mstcar}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
}
\examples{
}
