% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstcar-package.R
\docType{package}
\name{mstcar}
\alias{mstcar}
\alias{mstcar-package}
\title{mstcar}
\description{
Gibbs Sampler for MSTCAR Spatiotemporal Models
}
\author{
David DeLara <sfq1@cdc.gov>
}
\keyword{internal}
