% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncheart}
\alias{ncheart}
\title{Heart Attack Mortality Data in North Carolina}
\format{
a list with two array objects:

\describe{
  \item{Y}{death count (1-884)}
  \item{n}{population count (31-38243)}
}
}
\source{
\url{https://wonder.cdc.gov/cmf-icd9.html}
}
\usage{
ncheart
}
\description{
A dataset containing counts for Myocardial Infarction (ICD-9 code 410.0)
deaths in 100 North Carolina counties for individuals in 4 age groups across 10 years. This
dataset also contains the corresponding population counts.
}
\keyword{datasets}
