% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncnb}
\alias{ncnb}
\title{North Carolina Neighbor Data}
\format{
a list with two objects:

\describe{
  \item{neigh}{List of neighbor adjacency vectors}
  \item{num}{Vector of number of neighbors for each county}
}
}
\usage{
ncnb
}
\description{
A dataset containing U.S. Census TIGER shape data for North Carolina
}
\keyword{datasets}
