% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncshp}
\alias{ncshp}
\title{North Carolina Shapefile}
\format{
a data frame with 100 rows and 6 variables:

\describe{
  \item{STATEFP}{State FIPS code}
  \item{COUNTYFP}{County FIPS code}
  \item{COUNTYNS}{County GNIS code}
  \item{AFFGEOID}{Census Unique Identifier}
  \item{GEOID}{Census Unique Identifier, truncated}
  \item{NAME}{County Name}
  \item{LSAD}{County LSAD code}
  \item{ALAND}{amount of land in square meters}
  \item{AWATER}{amount of water in square meters}
  \item{geometry}{shape data for each county}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html}
}
\usage{
ncshp
}
\description{
A dataset containing U.S. Census TIGER shape data for North Carolina
}
\keyword{datasets}
