% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{samples}
\alias{samples}
\title{Gibbs Sampler Updates}
\usage{
samples(mod, n_iter, r = 100)
}
\arguments{
\item{mod}{a model object created with the \code{init()} function.}

\item{n_iter}{number of iterations to sample. Must be a multiple of 100 and be ≥ \code{r}.}

\item{r}{number of iterations to store in each batch. Must be a multiple of 100 and multiply evenly into \code{n_iter}.}
}
\description{
Gibbs Sampler Updates
}
\examples{
\dontrun{
samples(mod_nc, 6000)
samples(mod = mod_nc, n_iter = 1e4, r = 500)
}
}
