\name{mstcar-package}
\alias{mstcar-package}
\alias{mstcar}
\docType{package}
\title{
\packageTitle{mstcar}
}
\description{
\packageDescription{mstcar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mstcar}
\packageIndices{mstcar}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mstcar}

Maintainer: \packageMaintainer{mstcar}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
