% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncheart}
\alias{ncheart}
\title{Heart Attack Mortality Data in North Carolina}
\format{
a data frame with 100 rows and 6 variables:

\describe{
  \item{year}{year of death count (1979)}
  \item{agegrp}{age group of individuals counted (65-74)}
  \item{county}{name of county for event count}
  \item{FIPS}{FIPS code for corresponding county name}
  \item{deaths}{death count (1-129)}
  \item{population}{population count (400-21061)}
}
}
\source{
\url{https://wonder.cdc.gov/cmf-icd9.html}
}
\usage{
ncheart
}
\description{
A dataset containing counts for Myocardial Infarction (ICD-9 code 410.0)
deaths in North Carolina counties for individuals age 65-74 in 1979. This
dataset also contains the corresponding population counts and county FIPS codes.
The variables are as follows:
}
\keyword{datasets}
